% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparql.R
\name{arkg_clean}
\alias{arkg_clean}
\title{Limpiar URIs de ArKG}
\usage{
arkg_clean(x)
}
\arguments{
\item{x}{Vector de caracteres a limpiar.}
}
\value{
Vector de caracteres limpio.
}
\description{
Quita el prefijo \verb{https://arkg.cl/}, reemplaza guiones bajos por
espacios y decodifica caracteres URL-encoded.
}
\keyword{internal}
