% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{arkg_endpoint}
\alias{arkg_endpoint}
\title{Endpoint SPARQL de ArKG}
\usage{
arkg_endpoint()
}
\value{
Cadena de caracteres con la URL del endpoint.
}
\description{
Devuelve la URL del endpoint SPARQL de ArKG. Se puede sobrescribir con
la variable de entorno \code{ARKG_SPARQL_ENDPOINT}.
}
\examples{
arkg_endpoint()
}
