% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparql.R
\name{arkg_sparql}
\alias{arkg_sparql}
\title{Ejecutar una consulta SPARQL en ArKG}
\usage{
arkg_sparql(query, endpoint = arkg_endpoint())
}
\arguments{
\item{query}{Cadena de texto con la consulta SPARQL.}

\item{endpoint}{URL del endpoint (por defecto arkg_endpoint())..}
}
\value{
Un \code{tibble} para consultas SELECT, o \code{logical} para ASK.
}
\description{
Envia una consulta SPARQL al endpoint de ArKG (MillenniumDB) mediante
POST y devuelve los resultados como tibble (para SELECT) o logico (ASK).
}
\examples{
\dontrun{
arkg_sparql("SELECT * WHERE { ?s ?p ?o } LIMIT 5")
}
}
