% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{arkg_user_agent}
\alias{arkg_user_agent}
\title{User-agent del paquete arkg}
\usage{
arkg_user_agent()
}
\value{
Cadena de caracteres para la cabecera User-Agent.
}
\description{
User-agent del paquete arkg
}
\keyword{internal}
