% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparql.R
\name{sparql_to_tibble}
\alias{sparql_to_tibble}
\title{Convertir SPARQL Results JSON a tibble}
\usage{
sparql_to_tibble(parsed)
}
\arguments{
\item{parsed}{Lista resultante de parsear el JSON de respuesta.}
}
\value{
Un \code{tibble} (SELECT) o \code{logical} (ASK).
}
\description{
Implementa el formato "SPARQL 1.1 Query Results JSON".
}
\keyword{internal}
